requirements.txt
discord.py
Flask
nest_asyncio